+++ packages/platforms/react-native/lib/client.ts
@@ -1,4 +1,4 @@
-import { InMemoryQueue, createClient } from '@bugsnag/core-performance'
+import { createClient } from '@bugsnag/core-performance'
 import createFetchDeliveryFactory from '@bugsnag/delivery-fetch-performance'
 import { createXmlHttpRequestTracker } from '@bugsnag/request-tracker-performance'
 import { AppRegistry, AppState } from 'react-native'
@@ -6,26 +6,31 @@ import { FileSystem } from 'react-native-file-access'
 import { AppStartPlugin, NetworkRequestPlugin } from './auto-instrumentation'
 import createClock from './clock'
 import createSchema from './config'
-import idGenerator from './id-generator'
+import createIdGenerator from './id-generator'
 import NativeBugsnagPerformance from './native'
 import persistenceFactory from './persistence'
 import { platformExtensions } from './platform-extensions'
 import resourceAttributesSourceFactory from './resource-attributes-source'
+import createRetryQueueFactory from './retry-queue'
 import { createSpanAttributesSource } from './span-attributes-source'
+import createBrowserBackgroundingListener from './backgrounding-listener'
 
 const clock = createClock(performance)
 const appStartTime = clock.now()
 const deliveryFactory = createFetchDeliveryFactory(fetch, clock)
-const spanAttributesSource = createSpanAttributesSource(AppState)
+const spanAttributesSource = createSpanAttributesSource()
 const deviceInfo = NativeBugsnagPerformance ? NativeBugsnagPerformance.getDeviceInfo() : undefined
 const persistence = persistenceFactory(FileSystem, deviceInfo)
 const resourceAttributesSource = resourceAttributesSourceFactory(persistence, deviceInfo)
+const backgroundingListener = createBrowserBackgroundingListener(AppState)
 
 // React Native's fetch polyfill uses xhr under the hood, so we only track xhr requests
 const xhrRequestTracker = createXmlHttpRequestTracker(XMLHttpRequest, clock)
 
+const idGenerator = createIdGenerator(NativeBugsnagPerformance)
+
 const BugsnagPerformance = createClient({
-  backgroundingListener: { onStateChange: () => {} },
+  backgroundingListener,
   clock,
   deliveryFactory,
   idGenerator,
@@ -37,7 +42,7 @@ const BugsnagPerformance = createClient({
   resourceAttributesSource,
   schema: createSchema(),
   spanAttributesSource,
-  retryQueueFactory: (delivery, retryQueueMaxSize) => new InMemoryQueue(delivery, retryQueueMaxSize),
+  retryQueueFactory: createRetryQueueFactory(FileSystem),
   platformExtensions
 })